All IDs set to 8 bits

+++ fsm_keyboard.c
@@ -22,7 +22,7 @@
 /* State machine output or action functions */
 
 /* Other auxiliary functions */
-void fsm_keyboard_init(fsm_keyboard_t *p_fsm_keyboard, uint32_t keyboard_id)
+void fsm_keyboard_init(fsm_keyboard_t *p_fsm_keyboard, uint8_t keyboard_id)
 {
     // Initialize the FSM
     fsm_init(&p_fsm_keyboard->f, fsm_trans_keyboard);
@@ -32,7 +32,7 @@ void fsm_keyboard_init(fsm_keyboard_t *p_fsm_keyboard, uint32_t keyboard_id)
 }
 
 /* Public functions -----------------------------------------------------------*/
-fsm_keyboard_t *fsm_keyboard_new(uint32_t keyboard_id)
+fsm_keyboard_t *fsm_keyboard_new(uint8_t keyboard_id)
 {
     fsm_keyboard_t *p_fsm_keyboard = malloc(sizeof(fsm_keyboard_t)); /* Do malloc to reserve memory of all other FSM elements, although it is interpreted as fsm_t (the first element of the structure) */
     fsm_keyboard_init(p_fsm_keyboard, keyboard_id);                  /* Initialize the FSM */
